Simplify cli-handling functions
